Design Implementation

+++ src/com/uca/capas/dao/BookDAO.java
@@ -13,11 +13,11 @@ public interface BookDAO {
 	
 	public List<Book> findDetailed(String selector, String input) throws DataAccessException;
 	
-	public Integer countDetailed(String selector, String input) throws DataAccessException;
+	public BigInteger countDetailed(String selector, String input) throws DataAccessException;
 	
 	public List<Book> findDetailedCount(String selector, Integer input) throws DataAccessException;
 	
-	public Integer countDetailedNumber(String selector, Integer input) throws DataAccessException;
+	public BigInteger countDetailedNumber(String selector, Integer input) throws DataAccessException;
 	
 	public BigInteger countBooks() throws DataAccessException;
